fix errors

+++ src/app/cv/details-cv/details-cv.component.ts
@@ -39,7 +39,9 @@ export class DetailsCvComponent implements OnInit {
         },
       });
   }
-  deleteCv(cv: Cv) {
+  deleteCv(cv?: Cv|null) {
+    if(cv==null)
+      return;
     this.cvService.deleteCvById(cv.id).subscribe({
       next: () => {
         this.toastr.success(`${cv.name} supprimé avec succès`);
